+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased][unreleased]
+## Added
+- Implements array access for `Log` [#1](https://github.com/stevewest/changelog/issues/1)
+- Added name param to Release constructor.
 
 ## [0.1.0][0.1.0] - 2015-01-18
 ### Added